[REFACTOR] record 컴포넌트 및 훅 구조 분리
