+++ src/containers/record/star-log/StarAllComplete.tsx
