+++ src/containers/record/star-log/StarAnalysisDelayed.tsx
