+++ src/containers/record/star-log/StarImageUploader.tsx
